CleanUP API move messages to his own packet

+++ LightFallCorePlugin/src/main/java/de/lightfall/core/MessageProvider.java
@@ -1,14 +1,13 @@
 package de.lightfall.core;
 
 import co.aikar.commands.CommandManager;
-import co.aikar.locales.MessageKey;
 import co.aikar.locales.MessageKeyProvider;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.dao.DaoManager;
 import com.j256.ormlite.support.ConnectionSource;
 import com.j256.ormlite.table.TableUtils;
-import de.lightfall.core.api.IMessageKeyProvider;
-import de.lightfall.core.api.IMessageProvider;
+import de.lightfall.core.api.message.IMessageKeyProvider;
+import de.lightfall.core.api.message.IMessageProvider;
 import de.lightfall.core.models.MessageModel;
 import lombok.SneakyThrows;
 import net.md_5.bungee.api.ChatColor;
@@ -21,7 +20,7 @@ import java.util.logging.Logger;
 
 public class MessageProvider implements IMessageProvider {
 
-    private final Dao<MessageModel, Integer> messageDao;
+    private final Dao<MessageModel, Long> messageDao;
     private final CommandManager commandManager;
     private Logger logger;
